Removed DefaultSettingsSerializer & BaseOvservableJsonSettings
