+++ src/Files.App/Services/Settings/FileTagsSettingsService.cs
@@ -31,11 +31,12 @@ namespace Files.App.Services.Settings
 
 		public FileTagsSettingsService()
 		{
-			SettingsSerializer = new DefaultSettingsSerializer();
-			JsonSettingsDatabase = new CachingJsonSettingsDatabase(SettingsSerializer);
-
-			Initialize(Path.Combine(ApplicationData.Current.LocalFolder.Path,
-				Constants.LocalSettings.SettingsFolderName, Constants.LocalSettings.FileTagSettingsFileName));
+			JsonSettingsDatabase =
+				new DefaultJsonSettingsDatabase(
+					Path.Combine(
+						ApplicationData.Current.LocalFolder.Path,
+						Constants.LocalSettings.SettingsFolderName,
+						Constants.LocalSettings.FileTagSettingsFileName));
 		}
 
 		public IList<TagViewModel> FileTagList
