+++ src/Files.App/Utils/Serialization/DefaultJsonSettingsDatabase.cs
@@ -3,26 +3,35 @@
 
 using System.Collections.Concurrent;
 using System.Text.Json;
+using Win32PInvoke = Files.App.Helpers.NativeFileOperationsHelper;
 
 namespace Files.App.Utils.Serialization
 {
 	internal class DefaultJsonSettingsDatabase : IJsonSettingsDatabase
 	{
+		// Fields & Properties
+
+		private IDictionary<string, object?>? _settingsCache;
+
+		private string? _filePath;
+
 		public static readonly JsonSerializerOptions jsonSerializerOptions = new()
 		{
 			WriteIndented = true
 		};
 
-		protected ISettingsSerializer SettingsSerializer { get; }
+		// Constructor
 
-		public DefaultJsonSettingsDatabase(ISettingsSerializer settingsSerializer)
+		public DefaultJsonSettingsDatabase(string jsonFilePath)
 		{
-			SettingsSerializer = settingsSerializer;
+			CreateFile(jsonFilePath);
 		}
+
+		// Methods
 
 		protected IDictionary<string, object?>? GetFreshSettings()
 		{
-			string data = SettingsSerializer.ReadFromFile();
+			string data = ReadFromFile();
 
 			if (string.IsNullOrWhiteSpace(data))
 				data = "null";
@@ -51,43 +60,74 @@ namespace Files.App.Utils.Serialization
 		{
 			var jsonData = JsonSerializer.Serialize(data, jsonSerializerOptions);
 
-			return SettingsSerializer.WriteToFile(jsonData);
+			return WriteToFile(jsonData);
 		}
 
 		public virtual TValue? GetValue<TValue>(string key, TValue? defaultValue = default)
 		{
-			var data = GetFreshSettings();
-			if (data is null)
-				return defaultValue;
+			_settingsCache ??= GetFreshSettings();
 
-			if (data.TryGetValue(key, out var objVal))
+			if (_settingsCache is not null && _settingsCache.TryGetValue(key, out var objVal))
 			{
 				return GetValueFromObject<TValue>(objVal) ?? defaultValue;
 			}
 			else
 			{
-				SetValue(key, defaultValue);
+				if (_settingsCache is null)
+					return defaultValue;
+
+				if (SetValue(key, defaultValue))
+					_settingsCache.TryAdd(key, defaultValue);
+
 				return defaultValue;
 			}
 		}
 
 		public virtual bool SetValue<TValue>(string key, TValue? newValue)
 		{
-			var data = GetFreshSettings();
-			if (data is null)
+			_settingsCache ??= GetFreshSettings();
+
+			if (_settingsCache is null)
 				return false;
 
-			if (!data.TryAdd(key, newValue))
-				data[key] = newValue;
+			if (_settingsCache.TryAdd(key, newValue))
+				return SaveSettings(_settingsCache);
+			else
+				return UpdateValueInCache(_settingsCache[key]);
 
-			return SaveSettings(data);
+			bool UpdateValueInCache(object? value)
+			{
+				bool isDifferent;
+
+				if (newValue is IEnumerable enumerableNewValue && value is IEnumerable enumerableValue)
+				{
+					isDifferent = !enumerableValue.Cast<object>().SequenceEqual(enumerableNewValue.Cast<object>());
+				}
+				else
+				{
+					isDifferent = value != (object?)newValue;
+				}
+
+				if (isDifferent)
+				{
+					// Values are different, update the value and reload the cache.
+					_settingsCache[key] = newValue;
+
+					return SaveSettings(_settingsCache);
+				}
+				else
+				{
+					// The cache does not need to be updated, continue.
+					return false;
+				}
+			}
 		}
 
 		public virtual bool RemoveKey(string key)
 		{
-			var data = GetFreshSettings();
+			_settingsCache ??= GetFreshSettings();
 
-			return data.Remove(key) && SaveSettings(data);
+			return _settingsCache is not null && _settingsCache.Remove(key) && SaveSettings(_settingsCache);
 		}
 
 		public bool FlushSettings()
@@ -103,15 +143,18 @@ namespace Files.App.Utils.Serialization
 				// Try convert
 				var data = (IDictionary<string, object?>?)import;
 				if (data is null)
-				{
 					return false;
-				}
 
 				// Serialize
 				var serialized = JsonSerializer.Serialize(data, jsonSerializerOptions);
 
 				// Write to file
-				return SettingsSerializer.WriteToFile(serialized);
+				if (!WriteToFile(serialized))
+					return false;
+
+				_settingsCache = GetFreshSettings();
+
+				return true;
 			}
 			catch (Exception ex)
 			{
@@ -135,6 +178,48 @@ namespace Files.App.Utils.Serialization
 			}
 
 			return (TValue?)obj;
+		}
+
+		public bool CreateFile(string path)
+		{
+			var parentDir = SystemIO.Path.GetDirectoryName(path);
+			if (string.IsNullOrEmpty(parentDir))
+				return false;
+
+			Win32PInvoke.CreateDirectoryFromApp(parentDir, IntPtr.Zero);
+
+			var hFile = Win32PInvoke.CreateFileFromApp(
+				path,
+				Win32PInvoke.GENERIC_READ,
+				Win32PInvoke.FILE_SHARE_READ,
+				IntPtr.Zero,
+				Win32PInvoke.OPEN_ALWAYS,
+				(uint)Win32PInvoke.File_Attributes.BackupSemantics,
+				IntPtr.Zero);
+
+			if (hFile.IsHandleInvalid())
+				return false;
+
+			Win32PInvoke.CloseHandle(hFile);
+
+			_filePath = path;
+			return true;
+		}
+
+		public string ReadFromFile()
+		{
+			if (string.IsNullOrEmpty(_filePath))
+				throw new ArgumentNullException(nameof(_filePath));
+
+			return Win32PInvoke.ReadStringFromFile(_filePath);
+		}
+
+		public bool WriteToFile(string? text)
+		{
+			if (string.IsNullOrEmpty(_filePath))
+				throw new ArgumentNullException(nameof(_filePath));
+
+			return Win32PInvoke.WriteStringToFile(_filePath, text);
 		}
 	}
 }
