--- src/Files.Core/Data/Contracts/ISettingsSerializer.cs
@@ -1,14 +1,0 @@
-// Copyright (c) 2023 Files Community
-// Licensed under the MIT License. See the LICENSE.
-
-namespace Files.Core.Data.Contracts
-{
-	public interface ISettingsSerializer
-	{
-		bool CreateFile(string path);
-
-		string ReadFromFile();
-
-		bool WriteToFile(string? text);
-	}
-}
