--- src/Files.Core/Utils/Serialization/BaseObservableJsonSettings.cs
@@ -1,26 +1,0 @@
-// Copyright (c) 2023 Files Community
-// Licensed under the MIT License. See the LICENSE.
-
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
-
-namespace Files.App.Utils.Serialization
-{
-	public abstract class BaseObservableJsonSettings : BaseJsonSettings, INotifyPropertyChanged
-	{
-		public event PropertyChangedEventHandler? PropertyChanged;
-
-		protected override bool Set<TValue>(TValue? value, [CallerMemberName] string propertyName = "") where TValue : default
-		{
-			if (!base.Set<TValue>(value, propertyName))
-				return false;
-
-			OnPropertyChanged(propertyName);
-
-			return true;
-		}
-
-		protected void OnPropertyChanged([CallerMemberName] string propertyName = "")
-			=> PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-	}
-}
